backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-js-asset==3.1.2
djangorestframework==3.15.2
drf-yasg==1.21.10
django-filter==24.1
inflection==0.5.1
packaging==24.2
pillow==11.1.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
sqlparse==0.5.3
uritemplate==4.1.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9